refactor: split control plane tests
